requirements.txt
# Core
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pdfplumber>=0.10.0

# Gemini API
google-genai>=1.0.0
